0513-Find-Bottom-Left-Value/Solution_test.go
package leetcode

import (
	"github.com/duanbiaowu/leetcode/structures"
	"testing"
)

func Test_findBottomLeftValue(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			"test-1",
			args{nil},
			-1,
		},
		{
			"test-2",
			args{&TreeNode{Val: 1}},
			1,
		},
		{
			"test-3",
			args{structures.GenerateTreeNodesBySlice([]int{2, 1, 3})},
			1,
		},
		{
			"test-4",
			args{structures.GenerateTreeNodesBySlice([]int{1, 2, 3, 4, structures.NULL, 5, 6, structures.NULL, structures.NULL, 7})},
			7,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findBottomLeftValue(tt.args.root); got != tt.want {
				t.Errorf("findBottomLeftValue() = %v, want %v", got, tt.want)
			}
		})
	}
}
